chore(arch): clear deferred architecture items (READMEs, inlines, test ports)
Produced by an autonomous codex session, reviewed and committed here.

- Add 18 folder-boundary READMEs and drop the minFolderReadmeChildren
  relaxation from every config, so folder-readme-required is satisfied by real
  boundary docs rather than a raised threshold.
- Inline 8 trivial single-consumer sink files into their call sites, removing
  the no-trivial-sink-file waivers (the Nanoclaw host-substitution seam stays a
  documented waiver).
- Introduce server-owned test ports (test-utils/ports.ts) exposing database,
  readiness, and tracing projections; raw Kysely, OpenTelemetry, and PGlite
  handles stay internal, removing the vendor-type-leak waiver and its TRIAGE.
- Protocol #-subpath false-positive waivers now reference upstream
  chughtapan/safer-architecture-lsp#2.

pnpm arch:check clean: 0 findings, 17 waivers (down from 26).
