+++ TRIAGE-arch.md
@@ -1,8 +1,12 @@
 # Follow up on deferred safer-architecture findings
 
 `@chughtapan/safer-architecture-lsp` is enforced by `pnpm lint`. The initial
-adoption residual is clean, with two source-scoped decisions that intentionally
-remain follow-up work.
+adoption residual is clean, with one source-scoped decision that intentionally
+remains follow-up work.
+
+Protocol `#`-subpath boundary findings remain waived as analyzer false
+positives. Each directive references upstream
+`chughtapan/safer-architecture-lsp#2`.
 
 ## Deferred findings
 
@@ -13,16 +17,6 @@ remain follow-up work.
     is added. Split only along a stable RPC, callback, or notification boundary,
     while preserving the protocol layer DAG.
 
-- `no-public-vendor-type-leak` —
-  `packages/server/src/test-utils/index.ts`
-  - Reason: The test harness exposes live Kysely and OpenTelemetry handles used
-    by integration tooling; define server-owned test ports before removing
-    them.
-  - Follow-up: Introduce package-owned test-harness ports for the database,
-    server process, and span exporter, then migrate downstream integration
-    consumers and remove the waiver. This one directive covers two original
-    diagnostics.
-
 ## Config-relaxed shape heuristics
 
 Reaching the green gate raised several folder- and surface-shape budgets in the
@@ -30,10 +24,6 @@ per-package `safer-architecture.config.json` files rather than restructuring the
 code. These are deliberate, config-first deferrals, not permanent decisions —
 revisit them as the packages evolve:
 
-- `minFolderReadmeChildren` was raised (client 26, runtimes 14, server 14,
-  protocol 13), which suppresses `folder-readme-required` for the current tree.
-  Follow-up: add `README.md` boundary statements to the larger folders and lower
-  the threshold back toward the default (4).
 - `maxPublicExports` / `maxPublicReexports` / `minPublicFacadeModules` /
   `minExportedSiblingModules` / `maxSubpathExports` were raised to match each
   package's present curated surface. Follow-up: treat a future increase as a
