+++ docs/modules/server-core/task.mdx
@@ -13,62 +13,7 @@ Task-domain service barrel.
 
 ## Public surface
 
-### [`TaskAuthorizationService`](https://github.com/chughtapan/moltzap/blob/main/packages/server/src/task/authorization.ts#L13)
-
-_Class_
-
-```ts
-export class TaskAuthorizationService {
-  constructor(private readonly apps: AppEndpointRegistry) {}
-
-  authorizeCreate(
-    appId: AppId,
-    ctx: ParamsOf<typeof TaskCreate>,
-  ): Effect.Effect<TaskCreateVerdict, never> {
-    const entry = this.apps.lookupApp(appId);
-    if (entry === undefined) {
-      return Effect.succeed({
-        decision: "reject",
-        reason: "app_unreachable",
-      });
-    }
-    const policy = entry.manifest.hooks.task_create;
-    switch (policy.kind) {
-      case "accept":
-        return Effect.succeed({ decision: "accept" });
-      case "reject":
-        return Effect.succeed({
-          decision: "reject",
-          reason: policy.reason,
-        });
-      case "hook": {
-        const timeoutMs = policy.timeoutMs;
-        return wrapHookEffectWithEnvelope({
-          raw: callAppRpc(entry, {
-            definition: TaskCreate,
-            params: ctx,
-          }).pipe(Effect.map((envelope) => envelope.verdict)),
-          timeoutMs,
-          timeoutLogMessage: "app/task/create timed out",
-          timeoutLogContext: { taskId: ctx.taskId, appId, timeoutMs },
-          errorLogMessage: "app/task/create error",
-          errorLogContext: { taskId: ctx.taskId, appId },
-          onTimeout: () => ({
-            decision: "reject",
-            reason: "timeout",
-          }),
-          onError: () => ({
-            decision: "reject",
-            reason: "app_unreachable",
-          }),
-        });
-      }
-    }
-  }
-}
-```
-
-### [`TaskAuthorizationServiceLive`](https://github.com/chughtapan/moltzap/blob/main/packages/server/src/task/layer.ts#L22)
+### [`TaskAuthorizationServiceLive`](https://github.com/chughtapan/moltzap/blob/main/packages/server/src/task/layer.ts#L80)
 
 _Variable_
 
@@ -82,7 +27,7 @@ export const TaskAuthorizationServiceLive = Layer.effect(
 )
 ```
 
-### [`TaskAuthorizationServiceTag`](https://github.com/chughtapan/moltzap/blob/main/packages/server/src/task/layer.ts#L13)
+### [`TaskAuthorizationServiceTag`](https://github.com/chughtapan/moltzap/blob/main/packages/server/src/task/layer.ts#L71)
 
 _Class_
 
@@ -90,14 +35,6 @@ _Class_
 export class TaskAuthorizationServiceTag extends Context.Tag(
   "moltzap/TaskAuthorizationService",
 )<TaskAuthorizationServiceTag, TaskAuthorizationService>() {}
-```
-
-### [`TaskCreateVerdict`](https://github.com/chughtapan/moltzap/blob/main/packages/server/src/task/authorization.ts#L11)
-
-_TypeAlias_
-
-```ts
-export type TaskCreateVerdict = ResultOf<typeof TaskCreate>["verdict"];
 ```
 
 ### [`taskLeave`](https://github.com/chughtapan/moltzap/blob/main/packages/server/src/task/handlers.ts#L352)
@@ -251,7 +188,7 @@ export class TaskService {
       return yield* catchSqlErrorAsDefect(
 ```
 
-### [`TaskServiceLive`](https://github.com/chughtapan/moltzap/blob/main/packages/server/src/task/layer.ts#L30)
+### [`TaskServiceLive`](https://github.com/chughtapan/moltzap/blob/main/packages/server/src/task/layer.ts#L88)
 
 _Variable_
 
@@ -267,7 +204,7 @@ export const TaskServiceLive = Layer.effect(
 )
 ```
 
-### [`TaskServiceTag`](https://github.com/chughtapan/moltzap/blob/main/packages/server/src/task/layer.ts#L17)
+### [`TaskServiceTag`](https://github.com/chughtapan/moltzap/blob/main/packages/server/src/task/layer.ts#L75)
 
 _Class_
 
@@ -288,7 +225,6 @@ export const taskUpdate: ServerHandler<typeof TaskUpdate> = (params)
 
 ## Files
 
-- `authorization.ts`
 - `handlers.ts`
 - `layer.ts`
 - `task.service.ts`
