+++ packages/client/src/channel-base/README.md
@@ -1,0 +1,13 @@
+# Channel adapter primitives
+
+This folder is the runtime-neutral base layer shared by MoltZap channel
+adapters.
+
+- `lease.ts`, `lease-store.ts`, and `lease-guard.ts` project wire lease errors,
+  track per-host leases, and enforce single-use replies.
+- `format-cross-conv.ts` and `format-group-block.ts` render enriched context
+  using caller-selected markup.
+- `index.ts` is the curated `@moltzap/client/channel-base` surface.
+
+Runtime process integration and channel-specific policy stay in the channel
+packages; this boundary only owns reusable admission and formatting mechanics.
