--- packages/client/src/cli/commands/agents.ts
@@ -1,54 +1,0 @@
-import { Args, Command } from "@effect/cli";
-import { Effect } from "effect";
-import { LocalDaemonCommands } from "../../local-daemon-rpc.js";
-import { command, runHandler } from "../transport.js";
-import { logJson, logLines } from "../output.js";
-
-// safer-arch-ignore folder-explicit-api-required: the CLI entrypoint deliberately composes private one-command-per-file leaves; this folder is not a reusable API.
-// safer-arch-ignore no-trivial-sink-file: each CLI command owns its parser and handler in a focused leaf module consumed by the single CLI entrypoint.
-const listAgents = Command.make("list", {}, () =>
-  runHandler(
-    command(LocalDaemonCommands.AgentsList, {}).pipe(
-      Effect.flatMap(logJson),
-      Effect.asVoid,
-    ),
-  ),
-).pipe(Command.withDescription("List agents (default)"));
-
-const namesArg = Args.text({ name: "name" }).pipe(
-  Args.withDescription("Agent names to look up"),
-  Args.repeated,
-);
-
-const lookupAgents = Command.make("lookup", { names: namesArg }, ({ names }) =>
-  runHandler(
-    command(LocalDaemonCommands.AgentsSearch, { names }).pipe(
-      Effect.flatMap((result) => {
-        if (result.agents.length === 0) {
-          return Effect.log("No agents found.");
-        }
-        return logLines(
-          result.agents.map((agent) => {
-            let line = `Agent: ${agent.name}\n  ID: ${agent.id}\n  Status: ${agent.status}`;
-            if (agent.description)
-              line += `\n  Description: ${agent.description}`;
-            return `${line}\n`;
-          }),
-        );
-      }),
-      Effect.asVoid,
-    ),
-  ),
-).pipe(Command.withDescription("Look up agents by name"));
-
-/**
- * `moltzap agents [list|lookup]` — default (no subcommand) lists all agents,
- * `lookup` resolves one or more names.
- */
-export const agentsCommand = Command.make("agents", {}, () =>
-  // Bare `moltzap agents` with no subcommand defaults to listing.
-  listAgents.handler({}),
-).pipe(
-  Command.withDescription("List and look up agents on MoltZap"),
-  Command.withSubcommands([listAgents, lookupAgents]),
-);
