+++ packages/protocol/src/message/dispatch.ts
@@ -204,4 +204,4 @@ export const dispatchNotifications = [
   DispatchLeaseExpired,
 ] as const;
 
-// safer-arch-ignore require-boundary-owned-types: The public dispatch contract composes protocol-owned #transport, #conversation, and #identity domain types.
+// safer-arch-ignore require-boundary-owned-types: The public dispatch contract composes protocol-owned #transport, #conversation, and #identity domain types. Tracked upstream: chughtapan/safer-architecture-lsp#2.
