+++ packages/protocol/src/rpc.ts
@@ -7,8 +7,8 @@
  * descriptor-construction transport layer.
  */
 
-// safer-arch-ignore no-public-vendor-type-leak: This facade re-exports protocol's own #transport subpath, which is package-owned rather than a vendor boundary.
-// safer-arch-ignore require-boundary-owned-types: This facade deliberately exposes package-owned #transport RPC support types.
+// safer-arch-ignore no-public-vendor-type-leak: This facade re-exports protocol's own #transport subpath, which is package-owned rather than a vendor boundary. Tracked upstream: chughtapan/safer-architecture-lsp#2.
+// safer-arch-ignore require-boundary-owned-types: This facade deliberately exposes package-owned #transport RPC support types. Tracked upstream: chughtapan/safer-architecture-lsp#2.
 
 export type {
   CallErrorsOf,
