+++ packages/protocol/src/socket/catalog/index.ts
@@ -174,4 +174,4 @@ export const ReverseRpcGroup = makeRpcGroup([
   ...notificationDefinitions.map((definition) => definition.notificationRpc),
 ]);
 
-// safer-arch-ignore require-boundary-owned-types: The public socket catalog is the composition facade for protocol-owned identity, network, task, conversation, and message descriptor groups.
+// safer-arch-ignore require-boundary-owned-types: The public socket catalog is the composition facade for protocol-owned identity, network, task, conversation, and message descriptor groups. Tracked upstream: chughtapan/safer-architecture-lsp#2.
