+++ packages/server/safer-architecture.config.json
@@ -49,7 +49,6 @@
     }
   ],
   "minExportedSiblingModules": 7,
-  "minFolderReadmeChildren": 14,
   "folderChildCountOverrides": [
     {
       "folder": ".",
