+++ packages/server/src/__tests__/conformance/suite.test.ts
@@ -34,7 +34,7 @@ import {
 import {
   startCoreTestServer,
   stopCoreTestServer,
-} from "../../test-utils/server.js";
+} from "../../test-utils/index.js";
 
 const ENABLED_ENV_VALUE = "1";
 const DEFAULT_TOXIPROXY_URL = "http://127.0.0.1:8474";
