+++ packages/server/src/dispatch/admission.service.ts
@@ -7,6 +7,7 @@ import type { AgentId, AppId, UserId } from "@moltzap/protocol/identity";
 import type { ConversationId, MessageId } from "@moltzap/protocol/conversation";
 import type { ConnectionId } from "@moltzap/protocol/socket";
 import type { ParamsOf } from "@moltzap/protocol/rpc";
+import type { TaskId } from "@moltzap/protocol/task";
 import type { NetworkSendServiceTag } from "#network";
 import {
   callAppRpc,
@@ -15,10 +16,6 @@ import {
 } from "#identity/apps";
 import type { Db } from "#db";
 import { catchSqlErrorAsDefect, takeFirstOption } from "#db";
-import {
-  lookupAppBoundForConversation,
-  type AppBoundConversationLookup,
-} from "./app-bound-conversation.js";
 import type {
   LeaseRegistry,
   LeaseVerdict,
@@ -26,6 +23,48 @@ import type {
 } from "./lease-registry.js";
 
 export type DispatchAuthorizeContext = ParamsOf<typeof DispatchAuthorize>;
+
+interface AppBoundConversationLookup {
+  readonly _tag: "AppBound";
+  readonly taskId: TaskId;
+  readonly appId: AppId;
+}
+
+/**
+ * Dispatch admission is only defined for app-bound, non-archived
+ * conversations. The success type has no non-app-bound arm, so downstream
+ * lease minting cannot accidentally handle one as a lease binding.
+ */
+function lookupAppBoundForConversation(
+  db: Db,
+  conversationId: ConversationId,
+): Effect.Effect<AppBoundConversationLookup, never, never> {
+  return catchSqlErrorAsDefect(
+    Effect.gen(function* () {
+      const rowOpt = yield* takeFirstOption(
+        db
+          .selectFrom("conversations")
+          .innerJoin("tasks", "tasks.id", "conversations.task_id")
+          .select(["tasks.id as task_id", "tasks.app_id"])
+          .where("conversations.id", "=", conversationId)
+          .where("conversations.archived_at", "is", null)
+          .where("tasks.app_id", "is not", null)
+          .limit(1),
+      );
+      if (Option.isNone(rowOpt) || rowOpt.value.app_id === null) {
+        return yield* Effect.dieMessage(
+          `agent/dispatch/request requires an app-bound conversation: ${conversationId}`,
+        );
+      }
+      const lookup: AppBoundConversationLookup = {
+        _tag: "AppBound",
+        taskId: rowOpt.value.task_id,
+        appId: rowOpt.value.app_id,
+      };
+      return lookup;
+    }).pipe(Effect.withSpan("lookupAppBoundForConversation")),
+  );
+}
 
 export type DispatchAdmissionResult =
   | {
